refactor(refactor): refactored codes to meet standards
