+++ tests/client/component/CreateUser.spec.js
@@ -5,7 +5,7 @@ import { mount } from 'enzyme';
 import { Provider } from 'react-redux';
 import React from 'react';
 import { CreateUser } from 
-'../../../client/src/components/CreateUser.jsx';
+'../../../client/src/components/user/CreateUser.jsx';
 import initialState from '../../../client/src/store/initialState';
 import configureStore from '../../../client/src/store/configureStore';
